backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
python-dotenv==1.0.1
pydantic==2.9.2
google-generativeai==0.8.3
google-api-core==2.24.0
tenacity==9.0.0
chromadb==0.5.18
pandas==2.2.3
weasyprint==62.3
jinja2==3.1.4
feedparser==6.0.11
httpx==0.27.2
backboard-sdk==1.5.3
numpy>=1.26.0
